OneOCR v2.0: change parameter to enum & rename classes to be consistent with windows SDK
