+++ ClientLibrary/lib/src/main/java/com/microsoft/projectoxford/vision/contract/TextRecognitionResult.java
@@ -32,37 +32,16 @@
 //
 package com.microsoft.projectoxford.vision.contract;
 
-import java.util.ArrayList;
 import java.util.List;
-public class TextLine {
+public class TextRecognitionResult {
 
-    private List<TextWord> words; //words in the line
+    private List<TextRecognitionLine> lines; //lines in recognition result
 
-    private String text; //text of text line
-
-    private ArrayList<Integer> boundingBox; //bouding box for whole text line, e.g. "boundingBox": [153,579,1310,550,1313,647,156,678].
-
-    public List<TextWord> getWords() {
-        return words;
+    public List<TextRecognitionLine> getLines() {
+        return lines;
     }
 
-    public void setWords(List<TextWord> words) {
-        this.words = words;
-    }
-
-    public String getText() {
-        return text;
-    }
-
-    public void setText(String text) {
-        this.text = text;
-    }
-
-    public ArrayList<Integer> getBoundingBox() {
-        return boundingBox;
-    }
-
-    public void setBoundingBox(ArrayList<Integer> boundingBox) {
-        this.boundingBox = boundingBox;
+    public void setLines(List<TextRecognitionLine> lines) {
+        this.lines = lines;
     }
 }
